Refactored action rule ui components into separate modules. Also addressed remaining PR 543 comments.
