+++ hasher-matcher-actioner/webapp/src/pages/settings/ActionRuleSettingsTab.jsx
@@ -4,15 +4,15 @@
 
 /* eslint-disable react/prop-types */
 
-import React, {useState} from 'react';
 import Button from 'react-bootstrap/Button';
 import Col from 'react-bootstrap/Col';
 import Container from 'react-bootstrap/Container';
-import Form from 'react-bootstrap/Form';
-import Modal from 'react-bootstrap/Modal';
+import React, {useState} from 'react';
 import Row from 'react-bootstrap/Row';
 import Table from 'react-bootstrap/Table';
 import Toast from 'react-bootstrap/Toast';
+import ActionRuleFormColumns from '../../components/settings/ActionRuleFormColumns';
+import ActionRulesTableRow from '../../components/settings/ActionRulesTableRow';
 import '../../styles/_settings.scss';
 
 const mockedActionRules = [
@@ -44,97 +44,17 @@ const actions = [
   },
 ];
 
-function ActionRuleFormColumns(props) {
-  const {
-    name,
-    mustHaveLabels,
-    mustNotHaveLabels,
-    action,
-    showErrors,
-    nameIsUnique,
-    oldName,
-    onChange,
-  } = props;
-
-  const actionOptions = actions.map(actn => (
-    <option key={actn.id} value={actn.id}>
-      {actn.name}
-    </option>
-  ));
-
-  return (
-    <>
-      <td>
-        <Form.Label>
-          Name
-          <span hidden={!showErrors || name}> (required)</span>
-        </Form.Label>
-        <Form.Control
-          type="text"
-          required
-          value={name}
-          onChange={e => onChange({name: e.target.value})}
-          isInvalid={showErrors && !name}
-        />
-        <Form.Text
-          hidden={!showErrors || nameIsUnique(name, oldName)}
-          className="text-danger">
-          An action rule&rsquo;s name must be unique.
-        </Form.Text>
-      </td>
-      <td>
-        <Form.Label>
-          Labeled As
-          <span hidden={!showErrors || mustHaveLabels}> (required)</span>
-        </Form.Label>
-        <Form.Control
-          as="textarea"
-          rows={4}
-          required
-          value={mustHaveLabels}
-          onChange={e => onChange({must_have_labels: e.target.value})}
-          isInvalid={showErrors && !mustHaveLabels}
-        />
-      </td>
-      <td>
-        <Form.Label>Not Labeled As</Form.Label>
-        <Form.Control
-          as="textarea"
-          rows={4}
-          value={mustNotHaveLabels}
-          onChange={e => onChange({must_not_have_labels: e.target.value})}
-        />
-      </td>
-      <td>
-        <Form.Label>
-          Action
-          <span hidden={!showErrors || action !== '0'}> (required)</span>
-        </Form.Label>
-        <Form.Control
-          as="select"
-          required
-          value={action}
-          onChange={e => onChange({action_id: e.target.value})}
-          isInvalid={showErrors && action === '0'}>
-          <option value="0" key="0">
-            Select an action...
-          </option>
-          {actionOptions}
-        </Form.Control>
-      </td>
-    </>
-  );
-}
+const defaultActionRule = {
+  name: '',
+  must_have_labels: '',
+  must_not_have_labels: '',
+  action_id: '0',
+};
 
 export default function ActionRuleSettingsTab() {
   const [actionRules, setActionRules] = useState(mockedActionRules);
   const [adding, setAdding] = useState(false);
-  const [newActionRule, setNewActionRule] = useState({
-    name: '',
-    must_have_labels: '',
-    must_not_have_labels: '',
-    action_id: '0',
-  });
+  const [newActionRule, setNewActionRule] = useState(defaultActionRule);
   const [showErrors, setShowErrors] = useState(false);
   const [showToast, setShowToast] = useState(false);
   const [toastMessage, setToastMessage] = useState('');
@@ -169,12 +89,7 @@ export default function ActionRuleSettingsTab() {
     actionRuleIsValid(actionRule, actionRules, oldName);
 
   const resetForm = () => {
-    setNewActionRule({
-      name: '',
-      must_have_labels: '',
-      must_not_have_labels: '',
-      action_id: '0',
-    });
+    setNewActionRule(defaultActionRule);
   };
 
   const addActionRule = actionRule => {
@@ -210,6 +125,7 @@ export default function ActionRuleSettingsTab() {
   const actionRulesTableRows = actionRules.map(actionRule => (
     <ActionRulesTableRow
       key={actionRule.name}
+      actions={actions}
       name={actionRule.name}
       mustHaveLabels={actionRule.must_have_labels}
       mustNotHaveLabels={actionRule.must_not_have_labels}
@@ -241,9 +157,7 @@ export default function ActionRuleSettingsTab() {
                   <th>
                     <Button
                       className="table-action-button"
-                      onClick={() => {
-                        setAdding(true);
-                      }}>
+                      onClick={() => setAdding(true)}>
                       <ion-icon name="add" size="large" />
                     </Button>
                   </th>
@@ -261,7 +175,6 @@ export default function ActionRuleSettingsTab() {
                       className="mb-2 table-action-button"
                       onClick={() => {
                         setShowErrors(false);
-
                         if (actionRuleIsValid(newActionRule, actionRules)) {
                           addActionRule(newActionRule);
                           resetForm();
@@ -295,6 +208,7 @@ export default function ActionRuleSettingsTab() {
                     </Button>
                   </td>
                   <ActionRuleFormColumns
+                    actions={actions}
                     name={newActionRule.name}
                     mustHaveLabels={newActionRule.must_have_labels}
                     mustNotHaveLabels={newActionRule.must_not_have_labels}
@@ -320,149 +234,6 @@ export default function ActionRuleSettingsTab() {
           </Toast>
         </div>
       </Container>
-    </>
-  );
-}
-
-function ActionRulesTableRow(props) {
-  const {
-    name,
-    mustHaveLabels,
-    mustNotHaveLabels,
-    actionId,
-    onDeleteActionRule,
-    onUpdateActionRule,
-    ruleIsValid,
-    nameIsUnique,
-  } = props;
-  const [editing, setEditing] = useState(false);
-  const [
-    showDeleteActionRuleConfirmation,
-    setShowDeleteActionRuleConfirmation,
-  ] = useState(false);
-  const [updatedActionRule, setUpdatedActionRule] = useState({
-    name,
-    must_have_labels: mustHaveLabels,
-    must_not_have_labels: mustNotHaveLabels,
-    action_id: actionId,
-  });
-  const [showErrors, setShowErrors] = useState(false);
-
-  const onUpdatedActionRuleChange = updatedField => {
-    setUpdatedActionRule({...updatedActionRule, ...updatedField});
-  };
-
-  const resetForm = () => {
-    setUpdatedActionRule({
-      name,
-      must_have_labels: mustHaveLabels,
-      must_not_have_labels: mustNotHaveLabels,
-      action_id: actionId,
-    });
-  };
-
-  return (
-    <>
-      <tr hidden={editing}>
-        <td>
-          <Button
-            className="mb-2 table-action-button"
-            onClick={() => {
-              setEditing(true);
-            }}>
-            <ion-icon name="pencil" size="large" className="ion-icon-white" />
-          </Button>{' '}
-          <Button
-            variant="secondary"
-            className="table-action-button"
-            onClick={() => setShowDeleteActionRuleConfirmation(true)}>
-            <ion-icon
-              name="trash-bin"
-              size="large"
-              className="ion-icon-white"
-            />
-          </Button>
-          <Modal
-            show={showDeleteActionRuleConfirmation}
-            onHide={() => setShowDeleteActionRuleConfirmation(false)}>
-            <Modal.Header closeButton>
-              <Modal.Title>Confirm Action Rule Delete</Modal.Title>
-            </Modal.Header>
-            <Modal.Body>
-              <p>
-                Please confirm you want to delete the action rule named{' '}
-                <strong>{name}</strong>.
-              </p>
-            </Modal.Body>
-            <Modal.Footer>
-              <Button
-                variant="secondary"
-                onClick={() => setShowDeleteActionRuleConfirmation(false)}>
-                Cancel
-              </Button>
-              <Button
-                variant="primary"
-                onClick={() => onDeleteActionRule(name)}>
-                Yes, Delete This Action Rule
-              </Button>
-            </Modal.Footer>
-          </Modal>
-        </td>
-        <td>{name}</td>
-        <td className="action-rule-classification-column">{mustHaveLabels}</td>
-        <td className="action-rule-classification-column">
-          {mustNotHaveLabels.length > 0 ? (
-            mustNotHaveLabels
-          ) : (
-            <span>&mdash;</span>
-          )}
-        </td>
-        <td>{actions.find(action => action.id === actionId).name}</td>
-      </tr>
-      <tr hidden={!editing}>
-        <td>
-          <Button
-            variant="outline-primary"
-            className="mb-2 table-action-button"
-            onClick={() => {
-              setShowErrors(false);
-
-              if (ruleIsValid(updatedActionRule, name)) {
-                onUpdateActionRule(name, updatedActionRule);
-                resetForm();
-                setEditing(false);
-              } else {
-                setShowErrors(true);
-              }
-            }}>
-            <ion-icon
-              name="checkmark"
-              size="large"
-              className="ion-icon-white"
-            />
-          </Button>{' '}
-          <Button
-            variant="outline-secondary"
-            className="table-action-button"
-            onClick={() => {
-              setShowErrors(false);
-              resetForm();
-              setEditing(false);
-            }}>
-            <ion-icon name="close" size="large" className="ion-icon-white" />
-          </Button>
-        </td>
-        <ActionRuleFormColumns
-          name={updatedActionRule.name}
-          mustHaveLabels={updatedActionRule.must_have_labels}
-          mustNotHaveLabels={updatedActionRule.must_not_have_labels}
-          action={updatedActionRule.action_id}
-          showErrors={showErrors}
-          nameIsUnique={nameIsUnique}
-          oldName={name}
-          onChange={onUpdatedActionRuleChange}
-        />
-      </tr>
     </>
   );
 }
